Modified uploads to create root volume
Minor tweak to create a docker volume with the entire root directory, not just uploads.  This will preserve all app.php and .env configuration changes upon container restart.  The container will fail to start if both the root directory and the mysql directory are not persistent volumes.  Either they both must be or neither.